+++ src/components/Sidebar/index.tsx
@@ -6,7 +6,7 @@ import { Link, useLocation } from 'react-router-dom'
 import { useRecoilState, useSetRecoilState } from 'recoil'
 import LogoImg from 'src/assets/chatbot.png'
 import { useSettings } from 'src/hooks'
-import { db } from 'src/models/db'
+import { db } from 'src/db'
 import { currConversationState } from 'src/stores/conversation'
 import { currProductState } from 'src/stores/global'
 import { Products } from 'src/types/global'
